more class fixes after making uml diagram

+++ MathWiz/MathWiz/Student.cs
@@ -8,21 +8,21 @@ namespace MathWiz
 {
     public class Student : User
     {
-        public short MasteryLevel { get; set; }
+        public int MasteryLevel { get; set; }
         public List<Grade> Grades { get; set; }
 
         //empty constructor
         public Student() : base() { }
 
         //normal constructor
-        public Student(string username, string firstName, string lastName, short masteryLevel, List<Grade> grades = null) : base(username, firstName, lastName)
+        public Student(string username, string firstName, string lastName, int masteryLevel, List<Grade> grades = null) : base(username, firstName, lastName)
         {
             this.MasteryLevel = masteryLevel;
             this.Grades = grades;
         }
 
         //full constructor
-        public Student(string username, string firstName, string lastName, string password, short masteryLevel, List<Grade> grades) : base(username, firstName, lastName, password)
+        public Student(string username, string firstName, string lastName, string password, int masteryLevel, List<Grade> grades) : base(username, firstName, lastName, password)
         {
             this.MasteryLevel = masteryLevel;
             this.Grades = grades;
